requirements.txt
APScheduler==3.10.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
# certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
Django==4.0.3
django-apscheduler==0.6.2
django-cors-headers==3.11.0
django-crontab==0.7.1
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.3.0
numpy==1.24.2
pandas==1.5.3
psycopg-binary==3.1.8
psycopg2==2.9.5
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.2.4
tzdata==2021.5
tzlocal==4.2
